UUID presentation change and summary endpoint changes
Changed how the UUID gets presented in logs. Also made several updates to the summary endpoint queries. This now supports ADD_COLUMNS to show a list of unique values in the foreign column based on the filter. Also improved the data_source count summary to show the counts for each combination of data sources.